ICT2-91 신고 삭제 수정

+++ src/main/java/com/nutrimate/nutrimatebackend/model/challenge/ChallengeChatDto.java
@@ -26,8 +26,8 @@ public class ChallengeChatDto {
 	private String messageId;
 	private int COUNT;
 	
-	private Long chatroomId;
-	private Long userId;
+	private int chatroomId;
+	private long userId;
 	private MessageType messageType;
 	private String chatMessage;
 	private String challengeNick;
